splits axelrod tournament tests into separate game, player, round_robin and tournament files
